+++ src/gfx/env_map_gent.cpp
@@ -281,7 +281,7 @@ static bool LoadTex(char * FileName, unsigned char scdata [lthei][ltwid][3]){
 	int bpp=8;
 	int format=0;
 	unsigned char * palette;
-	unsigned char * data = readImage (fp,bpp,format,*(unsigned int*)&sizeX,*(unsigned int*)&sizeY,palette,texTransform,true);
+	unsigned char * data = readImage (fp,bpp,format,*(unsigned long*)&sizeX,*(unsigned long*)&sizeY,palette,texTransform,true);
 	bpp/=8;
 	if (format&PNG_HAS_ALPHA) {
 	  bpp*=4;
